Code fixes to unblock real-model test runs
  - Fixed local checkpoint shape mismatch by preserving
    language head geometry in local transformer config:
    mlx_audio/tts/models/moss_tts/config.py:122.
  - Fixed bfloat16 repetition-penalty runtime crash in real
    generation: mlx_audio/tts/models/moss_tts/sampling.py:60.
  - Fixed stream smoke deadlock source by making playback
    opt-in (--play) instead of implicit for --stream:
    mlx_audio/tts/generate.py:164.
  - Added regression coverage: mlx_audio/tts/tests/
    test_moss_tts_local_runtime.py:121, mlx_audio/tts/tests/
    test_moss_tts_local_runtime.py:265.

+++ mlx_audio/tts/models/moss_tts/sampling.py
@@ -64,8 +64,10 @@ def apply_repetition_penalty(
 
     if penalty == 1.0:
         return logits
-    logits_np = np.array(logits, copy=True)
-    history_np = np.array(previous_tokens)
+    # bfloat16 -> NumPy conversion can fail through the default bridge; run
+    # math in float32 NumPy space and cast back to the original MLX dtype.
+    logits_np = np.array(logits.astype(mx.float32), copy=True)
+    history_np = np.array(previous_tokens.astype(mx.int32))
 
     if history_np.ndim == 1:
         history_np = history_np[None, :]
